Extract GameTimerManager from PacmanGame (Architecture Phase 6)
Move all timer fields (fruitTime, frightModeTime, ghostModeTime,
ghostModeSwitchPos, forcePenLeaveTime) and the four timer handlers
(fruit, ghost-mode, force-pen-leave, gameplay-mode) out of PacmanGame
into a new GameTimerManager class using the game-reference pattern.

PacmanGame.handleTimers() now delegates to gameTimerManager.handleTimers().
Tests for the timer expiry state machine live in GameTimerManagerTest,
which uses a StubGame subclass to avoid the Mockito-spy/constructor-ref
split that caused the four handleTimers_* tests to fail.
